PROJETS/PRO/EURKAI/MODULES/MODEL_EXECUTOR/requirements.txt
# MODEL_EXECUTOR — Dépendances

pyyaml>=6.0
python-dotenv>=1.0.0

# Providers (install only needed ones)
openai>=1.0.0
anthropic>=0.18.0
replicate>=0.20.0
# elevenlabs>=0.2.0
# google-generativeai>=0.3.0
